+++ src/app/routes/broker-ui/catalog-page/state/catalog-page-actions.ts
@@ -9,7 +9,7 @@ export namespace CatalogPage {
   export class Reset {
     static readonly type = `[${tag}] Reset`;
 
-    constructor(public initialFilters?: FilterBoxModel[]) {}
+    constructor(public initialConnectorEndpoints?: string[]) {}
   }
 
   export class NeedFetch {
